Allow 3D animation to run on GPU's other than GPU0
The device (potentially a CUDA device) wasn't being passed to the instantiation of the AdaBins InferenceHelper. Thus it was always picking the default -- typically "cuda:0".
This change passes the same device initialized in the main notebook (Which yeah, is hardcoded to "cuda:0".. but is really easy to change to something else).