Save the custom model path when saving settings for a batch
When you have a lot of checkpoints of a model you use it is helpful to have the path to the checkpoint saved in the settings so you can go back and create more generations for a batch. Renamed custom_path to custom_model_path to better mirror the model_path variable and saved setting.